+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.2] - 2023-11-29
+
+### Added
+- Get-DefaultThrottleLimit
+- Invoke-RelativityInstallStaging
+
+### Changed
+- Added credential requirements to New-RelativityInstance and Get-RelativityInstance
+
 ## [0.1.1] - 2023-11-26
 
 ### Added